+++ workspaces/liferay-aicontentwizard-workspace/client-extensions/liferay-aicontentwizard-etc-spring-boot/src/main/java/com/liferay/ai/content/wizard/schemas/Account.java
@@ -5,8 +5,6 @@
 
 package com.liferay.ai.content.wizard.schemas;
 
-import com.liferay.ai.content.wizard.enums.AccountType;
-
 import dev.langchain4j.model.output.structured.Description;
 
 import org.json.JSONObject;
@@ -52,5 +50,20 @@ public class Account {
 		"Liferay provides three account types: Business, Person, and Supplier"
 	)
 	public AccountType type;
+
+	public enum AccountType {
+
+		@Description(
+			"Business accounts are used in B2C or B2X sites, default option."
+		)
+		business,
+		@Description("Used based on the site type (i.e., B2B, or B2C)")
+		person,
+		@Description(
+			"For accounts that are catalog suppliers, can publish products on their behalf"
+		)
+		supplier
+
+	}
 
 }
